+++ dags/grobid_build_image.py
@@ -3,7 +3,7 @@ from airflow.models import DAG
 from sciencebeam_airflow.utils.airflow import add_dag_macro
 from sciencebeam_airflow.utils.container_operators import ContainerRunOperator
 
-from sciencebeam_dag_ids import ScienceBeamDagIds
+from sciencebeam_airflow.dags.dag_ids import ScienceBeamDagIds
 
 from sciencebeam_dag_utils import (
     get_default_args,
