+++ dags/sciencebeam_add_model_config.py
@@ -5,7 +5,7 @@ import logging
 from airflow.operators.python import PythonOperator
 from airflow.models import DAG, DagRun
 
-from sciencebeam_dag_ids import ScienceBeamDagIds
+from sciencebeam_airflow.dags.dag_ids import ScienceBeamDagIds
 
 from sciencebeam_dag_utils import (
     get_default_args,
