+++ dags/sciencebeam_autocut_convert_training_data.py
@@ -2,7 +2,7 @@ import os
 
 from airflow.models import DAG
 
-from sciencebeam_dag_ids import ScienceBeamDagIds
+from sciencebeam_airflow.dags.dag_ids import ScienceBeamDagIds
 
 from sciencebeam_dag_utils import (
     get_default_args
