+++ dags/sciencebeam_convert.py
@@ -14,7 +14,7 @@ from sciencebeam_airflow.utils.container_operators import (
     HelmDeleteOperator
 )
 
-from sciencebeam_dag_ids import ScienceBeamDagIds
+from sciencebeam_airflow.dags.dag_ids import ScienceBeamDagIds
 
 from sciencebeam_dag_utils import (
     get_default_args,
